Refactor pipeline modules into subpackage
