+++ src/ume/__init__.py
@@ -32,6 +32,7 @@ except Exception:  # pragma: no cover - optional import
     def generate_embedding(text: str) -> list[float]:
         raise ImportError("sentence-transformers is required to generate embeddings")
 
+
 __all__ = [
     "Event",
     "EventType",
