+++ src/ume/pipeline/__init__.py
@@ -1,0 +1,6 @@
+"""Streaming pipeline utilities."""
+
+from .privacy_agent import run_privacy_agent
+from .stream_processor import build_app
+
+__all__ = ["run_privacy_agent", "build_app"]
